// packages/channels/src/wework/types.ts
type WeWorkNativeReplyType = 'file' | 'voice' | 'video';
type WeWorkNativeMediaBody = {
  media_id: string;
};
// ...
export interface WeWorkHttpResponseBody {
  msgtype: 'text' | 'markdown' | 'stream' | WeWorkNativeReplyType;
  text?: {
    content: string;
  };
  markdown?: {
    content: string;
  };
  file?: {
    media_id: string;
  };
  voice?: WeWorkNativeMediaBody;
  video?: WeWorkNativeMediaBody;
  stream?: {
    id: string;
    finish: boolean;
    content: string;
    msg_item?: Array<{
      msgtype: 'image' | 'file';
      image?: {
        base64: string;
        md5: string;
      };
    }>;
  };
}